checkout

+++ frontend/src/components/Checkout.jsx
@@ -10,8 +10,10 @@ import {
   Stack,
   Text,
   VStack,
+  useToast,
 } from "@chakra-ui/react";
 import axios from "axios";
+import { useState } from "react";
 import { FaCreditCard, FaMapMarkerAlt } from "react-icons/fa";
 import { HiOutlineCurrencyDollar } from "react-icons/hi";
 import { MdDriveFileRenameOutline } from "react-icons/md";
@@ -21,15 +23,53 @@ const url = "https://sleepy-erin-sheep.cyclic.app/";
 export const Checkout = () => {
   let cartData = JSON.parse(localStorage.getItem("cartData")) || null;
   let userData = JSON.parse(localStorage.getItem("userData")) || null;
+  const toast = useToast();
+  const [billingState, setBillingState] = useState({
+    name: "",
+    address: "",
+    city: "",
+    state: "",
+    pin: "",
+    cardNumber: "",
+    expiredDate: "",
+    cvv: "",
+  });
+
   const handleCheckout = () => {
-    localStorage.removeItem("cartData");
-    axios.delete(`${url}cart/delete`, {
-      headers: {
-        Authorization: `Bearer ${userData?.token}`,
-        "Content-Type": "application/json",
-      },
-    });
-    window.location.href = "/";
+    if (
+      billingState.name === "" ||
+      billingState.address === "" ||
+      billingState.city === "" ||
+      billingState.state === "" ||
+      billingState.pin === "" ||
+      billingState.cardNumber === "" ||
+      billingState.expiredDate === "" ||
+      billingState.cvv === ""
+    ) {
+      toast({
+        title: "Fail.",
+        description: "Please fill all the fields",
+        status: "error",
+        duration: 9000,
+        isClosable: true,
+      });
+    } else {
+      toast({
+        description:
+          "Your order has been  placed.Thank you for shopping with us",
+        status: "success",
+        duration: 9000,
+        isClosable: true,
+      });
+      localStorage.removeItem("cartData");
+      axios.delete(`${url}cart/delete`, {
+        headers: {
+          Authorization: `Bearer ${userData?.token}`,
+          "Content-Type": "application/json",
+        },
+      });
+      window.location.href = "/";
+    }
   };
 
   return (
@@ -53,13 +93,33 @@ export const Checkout = () => {
                   <FormLabel>
                     <MdDriveFileRenameOutline /> Name
                   </FormLabel>
-                  <Input type="text" borderRadius="md" />
+                  <Input
+                    type="text"
+                    borderRadius="md"
+                    placeholder="Name"
+                    value={billingState.name}
+                    onChange={(e) =>
+                      setBillingState((prev) => {
+                        return { ...prev, name: e.target.value };
+                      })
+                    }
+                  />
                 </FormControl>
                 <FormControl mt="4">
                   <FormLabel>
                     <FaMapMarkerAlt /> Address
                   </FormLabel>
-                  <Input type="text" borderRadius="md" />
+                  <Input
+                    type="text"
+                    borderRadius="md"
+                    placeholder="Address"
+                    value={billingState.address}
+                    onChange={(e) =>
+                      setBillingState((prev) => {
+                        return { ...prev, address: e.target.value };
+                      })
+                    }
+                  />
                 </FormControl>
                 <Stack
                   direction={{ base: "column", md: "row" }}
@@ -70,19 +130,49 @@ export const Checkout = () => {
                     <FormLabel>
                       <FaMapMarkerAlt /> City
                     </FormLabel>
-                    <Input type="text" borderRadius="md" />
+                    <Input
+                      type="text"
+                      borderRadius="md"
+                      placeholder="City"
+                      value={billingState.city}
+                      onChange={(e) =>
+                        setBillingState((prev) => {
+                          return { ...prev, city: e.target.value };
+                        })
+                      }
+                    />
                   </FormControl>
                   <FormControl flex="1">
                     <FormLabel>
                       <FaMapMarkerAlt /> State
                     </FormLabel>
-                    <Input type="text" borderRadius="md" />
+                    <Input
+                      type="text"
+                      borderRadius="md"
+                      placeholder="State"
+                      value={billingState.state}
+                      onChange={(e) =>
+                        setBillingState((prev) => {
+                          return { ...prev, state: e.target.value };
+                        })
+                      }
+                    />
                   </FormControl>
                   <FormControl flex="1">
                     <FormLabel>
                       <FaMapMarkerAlt /> Pin Code
                     </FormLabel>
-                    <Input type="text" borderRadius="md" />
+                    <Input
+                      type="text"
+                      borderRadius="md"
+                      placeholder="Pin"
+                      value={billingState.pin}
+                      onChange={(e) =>
+                        setBillingState((prev) => {
+                          return { ...prev, pin: e.target.value };
+                        })
+                      }
+                    />
                   </FormControl>
                 </Stack>
               </Box>
@@ -92,7 +182,17 @@ export const Checkout = () => {
                   <FormLabel>
                     <FaCreditCard /> Card Number
                   </FormLabel>
-                  <Input type="text" borderRadius="md" />
+                  <Input
+                    type="text"
+                    borderRadius="md"
+                    placeholder="Card Number"
+                    value={billingState.cardNumber}
+                    onChange={(e) =>
+                      setBillingState((prev) => {
+                        return { ...prev, cardNumber: e.target.value };
+                      })
+                    }
+                  />
                 </FormControl>
                 <Stack
                   direction={{ base: "column", md: "row" }}
@@ -103,13 +203,33 @@ export const Checkout = () => {
                     <FormLabel>
                       <HiOutlineCurrencyDollar /> Expiration Date
                     </FormLabel>
-                    <Input type="text" borderRadius="md" />
+                    <Input
+                      type="text"
+                      borderRadius="md"
+                      placeholder="Expiration Date"
+                      value={billingState.expiredDate}
+                      onChange={(e) =>
+                        setBillingState((prev) => {
+                          return { ...prev, expiredDate: e.target.value };
+                        })
+                      }
+                    />
                   </FormControl>
                   <FormControl flex="1">
                     <FormLabel>
                       <FaCreditCard /> CVV
                     </FormLabel>
-                    <Input type="text" borderRadius="md" />
+                    <Input
+                      type="text"
+                      borderRadius="md"
+                      placeholder="CVV"
+                      value={billingState.cvv}
+                      onChange={(e) =>
+                        setBillingState((prev) => {
+                          return { ...prev, cvv: e.target.value };
+                        })
+                      }
+                    />
                   </FormControl>
                 </Stack>
               </Box>
